LPS-204924 Move package, *TestUtil.java should be in test.util
