+++ modules/apps/data-engine/data-engine-rest-test/src/testIntegration/java/com/liferay/data/engine/rest/resource/v2_0/test/util/content/type/test/util/ModelResourceActionTestUtil.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.data.engine.rest.resource.v2_0.test.util.content.type;
+package com.liferay.data.engine.rest.resource.v2_0.test.util.content.type.test.util;
 
 import com.liferay.portal.kernel.model.ResourceAction;
 import com.liferay.portal.kernel.security.permission.ResourceActions;
